requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
decouple==0.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
django-import-export==2.0.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
nltk==3.4.5
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
